requirements.txt
ujson
aiohttp
aiofiles
asyncrcon
mcstatus
apscheduler
ruamel.yaml
aiocqhttp
nonebot
